requirements.txt
python-dotenv>=1.0.0
openai>=1.3.0
python-multipart>=0.0.6
